code refactor

+++ src/main/java/br/com/rafael/forum/controller/TopicosController.java
@@ -5,7 +5,6 @@ import java.util.List;
 import java.util.Optional;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
-import org.apache.catalina.connector.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -17,8 +16,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
-import br.com.rafael.forum.controller.dto.DetalhesTopicoDTO;
-import br.com.rafael.forum.controller.dto.TopicoDTO;
+import br.com.rafael.forum.controller.dto.DetalhesDoTopicoDto;
+import br.com.rafael.forum.controller.dto.TopicoDto;
 import br.com.rafael.forum.controller.form.AtualizacaoTopicoForm;
 import br.com.rafael.forum.controller.form.TopicoForm;
 import br.com.rafael.forum.modelo.Topico;
@@ -26,64 +25,69 @@ import br.com.rafael.forum.repository.CursoRepository;
 import br.com.rafael.forum.repository.TopicoRepository;
 
 @RestController
-@RequestMapping(value = "/topicos")
+@RequestMapping("/topicos")
 public class TopicosController {
 
   @Autowired
   private TopicoRepository topicoRepository;
+
   @Autowired
   private CursoRepository cursoRepository;
 
-
   @GetMapping
-  public List<TopicoDTO> lista(String nomeCurso) {
-
+  public List<TopicoDto> lista(String nomeCurso) {
     if (nomeCurso == null) {
       List<Topico> topicos = topicoRepository.findAll();
-      return TopicoDTO.converter(topicos);
+      return TopicoDto.converter(topicos);
     }
-
     List<Topico> topicos = topicoRepository.findByCursoNome(nomeCurso);
-    return TopicoDTO.converter(topicos);
+    return TopicoDto.converter(topicos);
   }
 
-  @Transactional
   @PostMapping
-  public ResponseEntity<TopicoDTO> cadastrar(@RequestBody @Valid TopicoForm form,
+  @Transactional
+  public ResponseEntity<TopicoDto> cadastrar(@RequestBody @Valid TopicoForm form,
       UriComponentsBuilder uriBuilder) {
-
     Topico topico = form.converter(cursoRepository);
-
     topicoRepository.save(topico);
 
     URI uri = uriBuilder.path("/topicos/{id}").buildAndExpand(topico.getId()).toUri();
-    return ResponseEntity.created(uri).body(new TopicoDTO(topico));
+    return ResponseEntity.created(uri).body(new TopicoDto(topico));
   }
 
   @GetMapping("/{id}")
-  public ResponseEntity<DetalhesTopicoDTO> detalhar(@PathVariable Long id) {
+  public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
     Optional<Topico> topico = topicoRepository.findById(id);
-
-    if (topico.isPresent())
-      return ResponseEntity.ok(new DetalhesTopicoDTO(topico.get()));
+    if (topico.isPresent()) {
+      return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+    }
 
     return ResponseEntity.notFound().build();
   }
 
-  @Transactional
   @PutMapping("/{id}")
-  public ResponseEntity<TopicoDTO> atualizar(@PathVariable Long id,
-      @RequestBody @Valid AtualizacaoTopicoForm form) {
-
-    Topico topico = form.atualizar(id, topicoRepository);
-
-    return ResponseEntity.ok(new TopicoDTO(topico));
-  }
-
   @Transactional
-  @DeleteMapping("/{id}")
-  public ResponseEntity deletar(@PathVariable Long id) {
-    topicoRepository.deleteById(id);
-    return ResponseEntity.ok().build();
+  public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id,
+      @RequestBody @Valid AtualizacaoTopicoForm form) {
+    Optional<Topico> optional = topicoRepository.findById(id);
+    if (optional.isPresent()) {
+      Topico topico = form.atualizar(id, topicoRepository);
+      return ResponseEntity.ok(new TopicoDto(topico));
+    }
+
+    return ResponseEntity.notFound().build();
   }
+
+  @DeleteMapping("/{id}")
+  @Transactional
+  public ResponseEntity<?> remover(@PathVariable Long id) {
+    Optional<Topico> optional = topicoRepository.findById(id);
+    if (optional.isPresent()) {
+      topicoRepository.deleteById(id);
+      return ResponseEntity.ok().build();
+    }
+
+    return ResponseEntity.notFound().build();
+  }
+
 }
